+++ packages/mangler-utils/src/index.ts
@@ -2,7 +2,6 @@ import type { SimpleManglerOptions } from "./simple-mangler.class";
 
 import {
   ALL_ALPHANUMERIC_CHARS,
-  ALL_CHARS,
   ALL_LETTER_CHARS,
   ALL_LOWERCASE_CHARS,
   ALL_NUMBER_CHARS,
@@ -44,7 +43,6 @@ const SimpleManglerPlugin = initSimpleManglerPlugin({
 
 export {
   ALL_ALPHANUMERIC_CHARS,
-  ALL_CHARS,
   ALL_LETTER_CHARS,
   ALL_LOWERCASE_CHARS,
   ALL_NUMBER_CHARS,