+++ packages/mangler-utils/CHANGELOG.md
@@ -8,7 +8,7 @@ Versioning].
 
 ## [Unreleased]
 
-- _No changes yet_
+- BREAKING: Remove `ALL_CHARS` character set. ([#247])
 
 ## [0.1.26] - 2022-01-17
 
@@ -30,5 +30,6 @@ Versioning].
 [#159]: https://github.com/ericcornelissen/webmangler/pull/159
 [#182]: https://github.com/ericcornelissen/webmangler/pull/182
 [#210]: https://github.com/ericcornelissen/webmangler/pull/210
+[#247]: https://github.com/ericcornelissen/webmangler/pull/247
 [keep a changelog]: https://keepachangelog.com/en/1.0.0/ "Keep a CHANGELOG"
 [semantic versioning]: https://semver.org/spec/v2.0.0.html "Semantic versioning"